+++ NEWS.md
@@ -1,4 +1,5 @@
 2020-01-15
+* GW-153: Presets can now be shared again.
 * GW-282: The most recent TAF schemas are used now when starting the BE which should keep these up to date if any changes have been made
 
 2020-01-14